video_uploader/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
PyJWT==1.7.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
